requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Wsgidav==4.1.0
mysql-connector-python==8.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==20.1.0
waitress==2.1.2
cheroot==9.0.0
psutil==7.1.3